# api/openapi.v1.yaml
openapi: 3.0.0
info:
  title: Shop
  description: The shop API.
  contact:
    name: Boykiv Bogdan
    email: boikiv.bogdan@gmail.com
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api'
...
paths:
...
  /category:
    get:
      summary: Получить список категорий товаров
      tags:
        - Category
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
    post:
      summary: Создать категорию
      tags:
        - Category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
      security:
        - BearerAuth: []
    patch:
      summary: Изменить категорию
      tags:
        - Category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
      security:
        - BearerAuth: []
  /category/{categoryId}:
    delete:
      summary: Удалить категорию
      tags:
        - Category
      responses:
        '200':
          description: successful operation
      security:
        - BearerAuth: []
    parameters:
      - schema:
          type: integer
          format: int64
        name: categoryId
        in: path
        required: true
        description: ID заказа
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    Category:
      title: Категория товара
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CategoryList:
      title: Список категорий товаров
      type: array
      items:
        $ref: '#/components/schemas/Category'
...
    CreateCategoryRequest:
      title: Тело запроса для создания категории
      type: object
      required:
        - name
      properties:
        name:
          type: string
...